refactor(ai): split dashscope payload coercion
